+++ app/src/main/java/com/example/king_bob_nae/features/home/presentation/viewholder/UserHolder.kt
@@ -1,4 +1,4 @@
-package com.example.king_bob_nae.features.home.presenter.viewholder
+package com.example.king_bob_nae.features.home.presentation.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
 import com.example.king_bob_nae.databinding.HolderFriendBinding
